+++ Examples/AspNetExample/Models/ClassTHeader.cs
@@ -1,0 +1,10 @@
+﻿using PRTelegramBot.Attributes;
+
+namespace TestDI.Models
+{
+    [InlineCommand]
+    internal enum ClassTHeader
+    {
+        DefaultTestClass = 555
+    }
+}