+++ Examples/AspNetExample/Program.cs
@@ -1,7 +1,9 @@
+using AspNetExample.BotController;
 using AspNetExample.Services;
 using PRTelegramBot.Core;
 using PRTelegramBot.Extensions;
 using PRTelegramBot.Models.EventsArgs;
+using TestDI.Models;
 
 /****************************************************************************************
  * ######################################################################################
@@ -20,6 +22,7 @@ builder.Services.AddControllersWithViews();
 builder.Services.AddTransient<ServiceTransient>();
 builder.Services.AddScoped<ServiceScoped>();
 builder.Services.AddSingleton<ServiceSingleton>();
+builder.Services.AddTransient<BotInlineHandlerWithDependency>();
 builder.Services.AddBotHandlers();
 
 
@@ -58,9 +61,10 @@ app.MapControllerRoute(
 
 //Создание и запуск бота
 var serviceProvaider = app.Services.GetService<IServiceProvider>();
-var prBotInstance = new PRBotBuilder("Token")
+var prBotInstance = new PRBotBuilder("token")
     .SetClearUpdatesOnStart(true)
     .SetServiceProvider(serviceProvaider)
+    .AddInlineClassHandler(ClassTHeader.DefaultTestClass, typeof(BotInlineHandlerWithDependency))
     .Build();
 
 prBotInstance.Events.OnCommonLog += PrBotInstance_OnLogCommon;