+++ Examples/ConsoleExample/Examples/Commands/ExampleInlineConfirmation.cs
@@ -5,6 +5,7 @@ using PRTelegramBot.Models;
 using PRTelegramBot.Models.CallbackCommands;
 using PRTelegramBot.Models.Enums;
 using PRTelegramBot.Models.InlineButtons;
+using PRTelegramBot.Models.TCommands;
 using PRTelegramBot.Utils;
 using Telegram.Bot;
 using Telegram.Bot.Types;
@@ -22,7 +23,7 @@ namespace ConsoleExample.Examples.Commands
         public static async Task InlineConfirm(ITelegramBotClient botClient, Update update)
         {
             //Кнопка для которой нужно создать подтверждение.
-            var exampleInlineCallback = new InlineCallback<EntityTCommand<long>>("Кнопка с подтвержением", CustomTHeaderTwo.ExampleThree, new EntityTCommand<long>(3, ActionWithLastMessage.Delete));
+            var exampleInlineCallback = new InlineCallback<EntityTCommand<long>>("Кнопка с подтверждением", CustomTHeaderTwo.ExampleThree, new EntityTCommand<long>(3, ActionWithLastMessage.Delete));
             //Обертка кнопки.
             var exampleWithConfirmation = new InlineCallbackWithConfirmation(exampleInlineCallback, ActionWithLastMessage.Delete);
 
@@ -34,6 +35,29 @@ namespace ConsoleExample.Examples.Commands
             //Передача меню в настройки
             option.MenuInlineKeyboardMarkup = testMenu;
             string msg = "InlineCallback с подтверждением";
+            //Отправка сообщение с меню
+            await PRTelegramBot.Helpers.Message.Send(botClient, update, msg, option);
+        }
+
+        /// <summary>
+        /// Пример обработки inline класса.
+        /// </summary>
+        [ReplyMenuHandler("InlineClass")]
+        public static async Task InlineClass(ITelegramBotClient botClient, Update update)
+        {
+            var exampleInlineCallback = new InlineCallback<StringTCommand>("Test1", ClassTHeader.DefaultTestClass, new StringTCommand("Test1"));
+            var exampleInlineCallbackTwo = new InlineCallback<StringTCommand>("Test2", ClassTHeader.DefaultTestClass, new StringTCommand("Test2"));
+            var exampleInlineCallbackThree = new InlineCallback<StringTCommand>("Test3", ClassTHeader.DefaultTestClass, new StringTCommand("Test3"));
+
+            //Создание нового меню.
+            List<IInlineContent> menu = new() { exampleInlineCallback, exampleInlineCallbackTwo, exampleInlineCallbackThree };
+            var testMenu = MenuGenerator.InlineKeyboard(1, menu);
+            var option = new OptionMessage();
+
+            //Передача меню в настройки
+            option.MenuInlineKeyboardMarkup = testMenu;
+            string msg = "InlineClass";
+
             //Отправка сообщение с меню
             await PRTelegramBot.Helpers.Message.Send(botClient, update, msg, option);
         }