+++ Examples/ConsoleExample/Models/CommandHeaders/ClassTHeader.cs
@@ -1,0 +1,10 @@
+﻿using PRTelegramBot.Attributes;
+
+namespace ConsoleExample.Models.CommandHeaders
+{
+    [InlineCommand]
+    internal enum ClassTHeader
+    {
+        DefaultTestClass = 555
+    }
+}